streamText tests
